Move/add tests
